tests/submission/Infraware/Session History/contents/Session_History/Not Use/history_length.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>history_length</h3>
	</p>

	<li>Close current browser and open new browser window, then test history length.</li>

	<hr>

	<div id="log"></div>

	<script>

	var len;
	len = window.history.length;
	test(function() {
		assert_equals(len, 2, "length attribute is failed.");
	});

	</script>
 </body>
</html>
